Voronoi.Border

Voronoi.Delaunay

Voronoi.Edge

Voronoi.HalfEdge

Voronoi.Triangle

Voronoi.Vertex

Voronoi.HalfEdges

Voronoi.areas

Voronoi.circumcenter

Voronoi.compute_areas

Voronoi.distance

Voronoi.filter_internal_triangles

Voronoi.find_triangle

Voronoi.flip!

Voronoi.game_over

Voronoi.get_common_points

Voronoi.get_points

Voronoi.in_triangle

Voronoi.insert_point!

Voronoi.insert_point_no_flip!

Voronoi.intersect_ray_bbox

Voronoi.is_delaunay

Voronoi.is_in_box

Voronoi.is_inside

Voronoi.is_left

Voronoi.main

Voronoi.place_point!

Voronoi.polygon_area

Voronoi.recursive_flip!

Voronoi.rotate_right

Voronoi.sort_vertices_ccw!

Voronoi.voronoi

Voronoi.winner

Voronoi.Border — Type

mutable struct Border <: Edge

An outer border edge used for triangulation initialization. Edges enclose their triangle in anticlockwise order. Due to the double-linked references, attributes may be initialized as nothing.

Attributes

origin::Vertex The origin vertex of this directed edge.

next::Union{Edge,Nothing} The next edge of the triangle.

prev::Union{Edge,Nothing} The preceding edge of the triangle.

face::Union{Face,Nothing} The triangle face containing this edge.

Constructors

Border(o::Vertex) only sets the origin vertex. Other attributes are nothing.

Border(o,n,p,f) sets all attributes.

Voronoi.Delaunay — Type

mutable struct Delaunay

The main triangulation object.

Attributes

triangles::Set{Triangle} Set of triangles in the Delaunay triangulation.

Constructors

Delaunay() Constructs a pseudo-empty triangulation, containing only one large

triangle encompassing the entire polygon.

Functions

D::Delaunay + T::Triangle adds the triangle to the triangulation.

D::Delaunay - T::Triangle removes the triangle from the triangulation.

Voronoi.Edge — Type

abstract type Edge

Represents an edge in the Delaunay triangulation.

Subtypes

@Border Edge outside of polygon space, used for initialization.

@HalfEdge Edge created when adding new vertices.

Voronoi.HalfEdge — Type

mutable struct HalfEdge <: Edge

An edge created when adding new points to the triangulation. Each half-edge is only contained within a single triangle. Adjacent triangles contain a twin of this edge pointing in the opposite direction. Edges enclose their triangle in anticlockwise order. Due to the double-linked references, attributes may be initialized as nothing.

Attributes

origin::Vertex The origin vertex of this directed edge.

twin::Union{HalfEdge,Nothing} The twin half-edge.

next::Union{Edge,Nothing} The next edge of the triangle.

prev::Union{Edge,Nothing} The preceding edge of the triangle.

face::Union{Face,Nothing} The triangle face containing this edge.

Constructors

HalfEdge(o::Vertex) only sets the origin. Other attributes are nothing.

HalfEdge(o,t,n,p,f) sets all attributes.

Voronoi.Triangle — Type

mutable struct Triangle <: Face

A triangle of the triangulation. Convenience object to connect edges.

Attributes

edge::Union{Edge,Nothing} One of the edges of the triangle.

Constructors

Triangle(xy::Edge, yz::Edge, zx::Edge) constructs a triangle XYZ from given

edges XY, YZ and ZX. This connects their next, prev and face attributes.

Functions

T1::Triangle == T2::Triangle determines if the triangles share any edge.

Voronoi.Vertex — Type

struct Vertex

Represents a point in the plane, placed by a player.

Attributes

x::Float64 The x-coordinate of the vertex, must be in [0,1].

y::Float64 The y-coordinate of the vertex, must be in [0,1].

player::Union{Int, Nothing} The player number, must be in {1,2}. For testing purposes,

a nothing value is permitted.

Constructors

Vertex(x,y,p) Constructs a vertex with player=p.

Vertex(x,y) Constructs a vertex with player=nothing.

Functions

a::Vertex + b::Vertex adds the coordinates into a vertex with player=nothing.

a::Vertex - b::Vertex subtracts the coordinates into a vertex with player=nothing.

a::Vertex * b::Vertex calculates the dot product, xx + yy.

k::Float64 * b::Vertex scales the coordinates by a factor.

Base.round — Method

round(a::Vertex; digits=10, base=2)

Rounds the x- and y-coordinates of a vertex to the given precision using @round.

Voronoi.HalfEdges — Method

HalfEdges(a::Vertex,b::Vertex)::Tuple{HalfEdge, HalfEdge}

Using vertices a and b, constructs two twin half-edges AB and BA.

Voronoi.areas — Method

areas(V::Dict{Vertex, Vector{Vertex}})::Dict{Int, Float64}

Calculates the Areas of the players

Input: Voronoi dict Output: Dict with the area of the polygons belonging to indiviual players

Voronoi.circumcenter — Method

circumcenter(T::Triangle)::Vertex

Calculates circumcenter coordinates of a given triangle.

Voronoi.compute_areas — Method

compute_areas(state::GameState)

Returns a Dict mapping player numbers to their Voronoi area.

Voronoi.distance — Method

distance(a::Vertex, b::Vertex)::Float64

calculates the distance between two vertices

Voronoi.filter_internal_triangles — Method

filter_internal_triangles(D::Delaunay)::Set{Triangle}

Filter out the triangles that have one corner in common with the initial Triangle in Delauney

Voronoi.find_triangle — Method

find_triangle(p::Vertex, D::Delaunay)::Triangle

Finds the triangle containing the given vertex. The vertex is to the left of each directed edge and does not lie on any edge. If such a triangle does not exist, nothing is returned.

Voronoi.flip! — Method

flip!(ab::HalfEdge, D::Delaunay):Tuple{Delaunay, Edge, Edge}

Should be called when the following conditions are met:

AB is in triangle ABC,

BA is in triangle BAP,

P is inside circumcircle of ABC.

Then it replaces triangles ABC, BAP with APC, BCP. Returns updated Delaunay structure and new edges AP,PB.

Voronoi.game_over — Method

game_over(state::GameState)::Bool

Returns true if both players have placed all their points.

Voronoi.get_common_points — Method

get_common_points(T1::Triangle, T2::Triangle)::Union{Vector{Vertex},Nothing}

calculates the common corners of two triangles. Returns nothing if there are none

Voronoi.get_points — Method

get_points(T::Triangle)::Vector{Vertex}

returns the corners of a triangle in a vetctor

Voronoi.in_triangle — Method

in_triangle(p::Vertex, T::Triangle)::Bool

checks if the point p is a corner of the triangle T

Voronoi.insert_point! — Method

insert_point!(p::Vertex, D::Delaunay)::Delaunay

Calls insert_point_no_flip! to split triangle ABC containing p into ABP, BCP and CAP. Then calls recursive_flip! to re-satisfy the Delaunay condition.

Example

```julia julia> D = Delaunay();

julia> insert_point!(Vertex(0.6, 0.9), D) Tri{Border[(0.0, 3.0)->(0.0, -2.0)], [(0.0, -2.0)->(0.6, 0.9)], [(0.6, 0.9)->(0.0, 3.0)]} Tri{Border[(0.0, -2.0)->(3.0, 0.0)], [(3.0, 0.0)->(0.6, 0.9)], [(0.6, 0.9)->(0.0, -2.0)]} Tri{Border[(3.0, 0.0)->(0.0, 3.0)], [(0.0, 3.0)->(0.6, 0.9)], [(0.6, 0.9)->(3.0, 0.0)]}

Voronoi.insert_point_no_flip! — Method

insert_point_no_flip!(p::Vertex, D::Delaunay)::Tuple{Delaunay,Edge,Edge,Edge}

Inserts point p into triangulation D by constructing new triangles. If p is inside triangle ABC, it is replaced by ABP, BCP and BAP. The old edges AB, BC and CA are also returned such that a flip! can be executed.

Voronoi.intersect_ray_bbox — Method

intersect_ray_bbox(origin::Vertex, dir::Tuple{Float64,Float64}, bbox::Vector{Vertex})

Calculates the intersection of a ray starting in bounding box(bbox)in direction of dir with the edges of the box

Voronoi.is_delaunay — Method

is_delaunay(ab::Edge)::Bool

Checks if the Delaunay condition is upheld for this edge. Borders always fulfill the Delaunay condition. For a half-edge AB in triangle ABC, the point D of triangle BAD must not be in the circumcircle of ABC.

Voronoi.is_in_box — Method

is_in_box(p::Vertex, bbox::Vector{Vertex})::Bool

Checks if the given Vertex is inside or on the border ob bbox

Voronoi.is_inside — Method

is_inside(v::Vertex)

Determines if the vertex coordinates place it inside [0,1]².

Voronoi.is_left — Method

is_left(p::Vertex, a::Vertex, b::Vertex)::Bool

Uses dot product to determine if a point p is left of a directed edge e. If the point lies exactly on the edge, it is not considered left of the edge.

Voronoi.main — Method

main()

Prompt for number of turns, then create the main window.

Voronoi.place_point! — Method

place_point!(state::GameState, x::Float64, y::Float64)

Attempts to place a point for the current player at (x, y). Returns true if successful, false otherwise.

Voronoi.polygon_area — Method

polygon_area(verts::Vector{Tuple{Float64,Float64}})::Float64

Calculates the area of a Polygon with n corners with shoelace formula

Input: vector of tuples (x, y), where x and y are Float64, sorted counterclockwise. Output: Area of the Polygon

Voronoi.recursive_flip! — Method

recursive_flip!(ab::Edge, D::Delaunay)::Delaunay

Flips edge AB if it breaks the Delaunay condition. If flipped, recursively checks AP and PB.

Voronoi.rotate_right — Method

rotate_right(h::HalfEdge)::Tuple{Float64,Float64}

Takes a HalfEdge and returns a direction, that is HalfEdge rotated by 90° counterclockwise

Voronoi.sort_vertices_ccw! — Method

sort_vertices_ccw!(verts::Vector{Tuple{Float64,Float64}})

Sorts a list of 2D points in counterclockwise order around their centroid.

Input: vector of tuples (x, y) where x and y are Float64 Output: vector of tuples (x, y) sorted counterclockwise

Voronoi.voronoi — Method

voronoi(D::Delaunay)::Tuple{Dict{Vertex,Vector{Vertex}},Dict{Vertex,Vector{Vertex}}}

Gives the Voronoi to a Delaunay Outputs: a dict V where each center of a Voronoi polygon is mapped to its corners. a dict A where each corner of a Voronoipolygon is mapped to its connected corners

Voronoi.winner — Method

winner(state::GameState)

Returns the player number with the largest area, or nothing if tied.